Sort Servername and ServerSSL DataGroup record items
Problem: Records in ServerName and ServerSSL DataGroups are
toggled during verify interval which leads to posting AS3
declaration to BIG-IP though there is no change in monitored
resources.

Solution: Always sort the order of records being posted so that
controller post only when there is a change is resources being
monitored.

Bug: CONTCNTR-1806

Affected-branches: master